Make some date related things configurable
You can now supply a config.dateStrings and config.humanMonths.

the config.dateStrings should be a function, which given a timestamp, and an array of month names, should return a hash of date strings, that dateStr then can use to render a proper date giver a interval.

config.humanMonths, if given, replaces the name of months passed to dateStrings.